Scrolling Gallery
In the interest of keeping movement and things alive, I wanted to have
a moving gallery for my portfolio.

There’s a seemingly infinite flow of images. JS creates a duplicate of
the images to fill in when the first set of images goes off screen.
Then the animation starts over and the effect is seamless.

Additionally, there are two ways of interacting with the galleries. On
the one hand, you can toggle a box below which gives more details about
the thing. On the other, the image is presented as a simple light box.

The former is for things that need some explication or are impressive
to employers and analytical types. The latter is for things that are
just to be contemplated and enjoyed.