Work Section, Byzantine Wheels
A lot of big changes today. The majority of the work was done on the
work section. Maybe that’s why it’s called the work section…

In any case, there are icons for services I provide. There is a client
concern followed by my answer. There’s a nice fade in for the initial
one dropped in the desktop version. And there is a three column layout
for mobile for the mobile version.

Also, the header image is now a computer with my byzantine wheel
spinning over top of it. I’m looking forward to making the animations
an little more interactive, but really the basic things are all here.
Everything from here on out with be fine tuning.

Things are starting to look like they were made by a web developer.
They’re also looking like they were built like me. It’s good to see
yourself in something you’ve made.